requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<=3.13
psycopg2>=2.8.6,<2.9
django-filter>=21.1,<25
drf-spectacular>=0.15.1,<0.16
uwsgi>=2.0.19,<2.1
requests>=2.26.0,<3